Chaange type of packet number length parameters

+++ iocore/net/quic/QUICHandshakeProtocol.h
@@ -59,10 +59,10 @@ public:
                        uint64_t pkt_num, const uint8_t *ad, size_t ad_len, QUICKeyPhase phase) const         = 0;
   virtual bool decrypt(uint8_t *plain, size_t &plain_len, size_t max_plain_len, const uint8_t *cipher, size_t cipher_len,
                        uint64_t pkt_num, const uint8_t *ad, size_t ad_len, QUICKeyPhase phase) const         = 0;
-  virtual bool encrypt_pn(uint8_t *protected_pn, size_t &protected_pn_len, const uint8_t *unprotected_pn, size_t unprotected_pn_len,
-                          const uint8_t *sample, QUICKeyPhase phase) const                                   = 0;
-  virtual bool decrypt_pn(uint8_t *unprotected_pn, size_t &unprotected_pn_len, const uint8_t *protected_pn, size_t protected_pn_len,
-                          const uint8_t *sample, QUICKeyPhase phase) const                                   = 0;
+  virtual bool encrypt_pn(uint8_t *protected_pn, uint8_t &protected_pn_len, const uint8_t *unprotected_pn,
+                          uint8_t unprotected_pn_len, const uint8_t *sample, QUICKeyPhase phase) const       = 0;
+  virtual bool decrypt_pn(uint8_t *unprotected_pn, uint8_t &unprotected_pn_len, const uint8_t *protected_pn,
+                          uint8_t protected_pn_len, const uint8_t *sample, QUICKeyPhase phase) const         = 0;
   virtual QUICHandshakeMsgType msg_type() const;
 
 protected:
